Docs, organizing. Added EvaluationContext.
Restructured a bit of how the expression evaluation flows through operators. Operands are evaluated before being passed to operator functions so the context doesn't need to be passed along.

EvaluationContext is what now holds the roller through the flow of the Evaluate() calls. I didn't want to add a parameter to evaluate, but this seemed like the most elegant solution to getting a roller where it needed to go. The class exists to allow for future context should it need to be added.

Tests will hang right now, because now that the proper rollers are being used, a Max Roller on exploding dice... yeah that rolls forever.